+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -5,6 +5,7 @@ package org.plumelib.util;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.BitSet;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -112,7 +113,7 @@ public final class CollectionsPlume {
    * @param <T> type of elements of the list
    * @param l a list to remove duplicates from
    * @return a copy of the list with duplicates removed
-   * @deprecated use {@link withoutDuplicates}
+   * @deprecated use {@link withoutDuplicates} or {@link withoutDuplicatesComparable}
    */
   @Deprecated // 2021-03-28
   public static <T> List<T> removeDuplicates(List<T> l) {
@@ -122,14 +123,37 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns a list with the same contents as its argument, but without duplicates. May return its
+   * Returns a copy of the list with duplicates removed. Retains the original order. May return its
    * argument if its argument has no duplicates, but is not guaranteed to do so.
+   *
+   * <p>If the element type implements {@link Comparable}, use {@link #withoutDuplicatesComparable}.
    *
    * @param <T> the type of elements in {@code values}
    * @param values a list of values
    * @return the values, with duplicates removed
    */
-  public static <T extends Comparable<T>> List<T> withoutDuplicates(List<T> values) {
+  public static <T> List<T> withoutDuplicates(List<T> values) {
+    HashSet<T> hs = new LinkedHashSet<>(values);
+    if (values.size() == hs.size()) {
+      return values;
+    } else {
+      return new ArrayList<>(hs);
+    }
+  }
+
+  /**
+   * Returns a list with the same contents as its argument, but without duplicates. May return its
+   * argument if its argument has no duplicates, but is not guaranteed to do so.
+   *
+   * <p>This is like {@link #withoutDuplicates}, but requires the list elements to implement {@link
+   * Comparable}, and thus can be more efficient. Also, this does not retain the original order; the
+   * result is sorted.
+   *
+   * @param <T> the type of elements in {@code values}
+   * @param values a list of values
+   * @return the values, with duplicates removed
+   */
+  public static <T extends Comparable<T>> List<T> withoutDuplicatesComparable(List<T> values) {
     // This adds O(n) time cost, and has the benefit of sometimes avoiding allocating a TreeSet.
     if (isSortedNoDuplicates(values)) {
       return values;
@@ -784,7 +808,9 @@ public final class CollectionsPlume {
 
     @Override
     public T next(@GuardSatisfied MergedIterator<T> this) {
-      hasNext(); // for side effect
+      if (!hasNext()) {
+        throw new NoSuchElementException();
+      }
       return current.next();
     }
 
@@ -1112,6 +1138,41 @@ public final class CollectionsPlume {
     return theKeys;
   }
 
+  /**
+   * Given an expected number of elements, returns the capacity that should be passed to a HashMap
+   * or HashSet constructor, so that the set or map will not resize.
+   *
+   * @param numElements the maximum expected number of elements in the map or set
+   * @return the initial capacity to pass to a HashMap or HashSet constructor
+   */
+  public static int mapCapacity(int numElements) {
+    // Equivalent to: (int) (numElements / 0.75) + 1
+    // where 0.75 is the default load factor.
+    return (numElements * 4 / 3) + 1;
+  }
+
+  /**
+   * Given an expected number of elements, returns the capacity that should be passed to a HashMap
+   * or HashSet constructor, so that the set or map will not resize.
+   *
+   * @param c a collection whose size is the maximum expected number of elements in the map or set
+   * @return the initial capacity to pass to a HashMap or HashSet constructor
+   */
+  public static int mapCapacity(Collection<?> c) {
+    return mapCapacity(c.size());
+  }
+
+  /**
+   * Given an expected number of elements, returns the capacity that should be passed to a HashMap
+   * or HashSet constructor, so that the set or map will not resize.
+   *
+   * @param m a map whose size is the maximum expected number of elements in the map or set
+   * @return the initial capacity to pass to a HashMap or HashSet constructor
+   */
+  public static int mapCapacity(Map<?, ?> m) {
+    return mapCapacity(m.size());
+  }
+
   ///////////////////////////////////////////////////////////////////////////
   /// Set
   ///
@@ -1134,5 +1195,108 @@ public final class CollectionsPlume {
       }
     }
     return null;
+  }
+
+  ///////////////////////////////////////////////////////////////////////////
+  /// BitSet
+  ///
+
+  /**
+   * Returns true if the cardinality of the intersection of the two BitSets is at least the given
+   * value.
+   *
+   * @param a the first BitSet to intersect
+   * @param b the second BitSet to intersect
+   * @param i the cardinality bound
+   * @return true iff size(a intersect b) &ge; i
+   */
+  @SuppressWarnings({"allcheckers:purity", "lock"}) // side effect to local state (BitSet)
+  @Pure
+  public static boolean intersectionCardinalityAtLeast(BitSet a, BitSet b, @NonNegative int i) {
+    // Here are three implementation strategies to determine the
+    // cardinality of the intersection:
+    // 1. a.clone().and(b).cardinality()
+    // 2. do the above, but copy only a subset of the bits initially -- enough
+    //    that it should exceed the given number -- and if that fails, do the
+    //    whole thing.  Unfortunately, bits.get(int, int) isn't optimized
+    //    for the case where the indices line up, so I'm not sure at what
+    //    point this approach begins to dominate #1.
+    // 3. iterate through both sets with nextSetBit()
+
+    int size = Math.min(a.length(), b.length());
+    if (size > 10 * i) {
+      // The size is more than 10 times the limit.  So first try processing
+      // just a subset of the bits (4 times the limit).
+      BitSet intersection = a.get(0, 4 * i);
+      intersection.and(b);
+      if (intersection.cardinality() >= i) {
+        return true;
+      }
+    }
+    return (intersectionCardinality(a, b) >= i);
+  }
+
+  /**
+   * Returns true if the cardinality of the intersection of the three BitSets is at least the given
+   * value.
+   *
+   * @param a the first BitSet to intersect
+   * @param b the second BitSet to intersect
+   * @param c the third BitSet to intersect
+   * @param i the cardinality bound
+   * @return true iff size(a intersect b intersect c) &ge; i
+   */
+  @SuppressWarnings({"allcheckers:purity", "lock"}) // side effect to local state (BitSet)
+  @Pure
+  public static boolean intersectionCardinalityAtLeast(
+      BitSet a, BitSet b, BitSet c, @NonNegative int i) {
+    // See comments in intersectionCardinalityAtLeast(BitSet, BitSet, int).
+    // This is a copy of that.
+
+    int size = Math.min(a.length(), b.length());
+    size = Math.min(size, c.length());
+    if (size > 10 * i) {
+      // The size is more than 10 times the limit.  So first try processing
+      // just a subset of the bits (4 times the limit).
+      BitSet intersection = a.get(0, 4 * i);
+      intersection.and(b);
+      intersection.and(c);
+      if (intersection.cardinality() >= i) {
+        return true;
+      }
+    }
+    return (intersectionCardinality(a, b, c) >= i);
+  }
+
+  /**
+   * Returns the cardinality of the intersection of the two BitSets.
+   *
+   * @param a the first BitSet to intersect
+   * @param b the second BitSet to intersect
+   * @return size(a intersect b)
+   */
+  @SuppressWarnings({"allcheckers:purity", "lock"}) // side effect to local state (BitSet)
+  @Pure
+  public static int intersectionCardinality(BitSet a, BitSet b) {
+    BitSet intersection = (BitSet) a.clone();
+    intersection.and(b);
+    return intersection.cardinality();
+  }
+
+  /**
+   * Returns the cardinality of the intersection of the three BitSets.
+   *
+   * @param a the first BitSet to intersect
+   * @param b the second BitSet to intersect
+   * @param c the third BitSet to intersect
+   * @return size(a intersect b intersect c)
+   */
+  @SuppressWarnings({"allcheckers:purity", "lock"}) // side effect to local state (BitSet)
+  @Pure
+  public static int intersectionCardinality(BitSet a, BitSet b, BitSet c) {
+    BitSet intersection = (BitSet) a.clone();
+    intersection.and(b);
+    intersection.and(c);
+    return intersection.cardinality();
   }
 }