+++ src/test/java/org/plumelib/util/CollectionsPlumeTest.java
@@ -232,8 +232,8 @@ public final class CollectionsPlumeTest {
           }
         }
         @SuppressWarnings({
-          "lowerbound:assignment.type.incompatible",
-          "index:assignment.type.incompatible",
+          "lowerbound:assignment",
+          "index:assignment",
           "value"
         }) // The IotaIterator only contains indexes for totals.length, and since chosen's
         // elements are selected randomly from the IotaIterator, all of its elements are
@@ -273,6 +273,16 @@ public final class CollectionsPlumeTest {
   List<Integer> l1123 = Arrays.asList(1, 1, 2, 3);
   List<Integer> l1233 = Arrays.asList(1, 2, 3, 3);
 
+  Object object1 = new Object();
+  Object object2 = new Object();
+  Object object3 = new Object();
+
+  List<Object> lo123 = Arrays.asList(object1, object2, object3);
+  List<Object> lo123123 = Arrays.asList(object1, object2, object3, object1, object2, object3);
+  List<Object> lo12223 = Arrays.asList(object1, object2, object2, object2, object3);
+  List<Object> lo1123 = Arrays.asList(object1, object1, object2, object3);
+  List<Object> lo1233 = Arrays.asList(object1, object2, object3, object3);
+
   @SuppressWarnings("JdkObsolete") // test of List that does not implement RandomAccess
   @Test
   public void test_hasDuplicates() {
@@ -301,6 +311,26 @@ public final class CollectionsPlumeTest {
     assertEquals(l123, CollectionsPlume.withoutDuplicates(l12223));
     assertEquals(l123, CollectionsPlume.withoutDuplicates(l1123));
     assertEquals(l123, CollectionsPlume.withoutDuplicates(l1233));
+
+    assertEquals(lo123, CollectionsPlume.withoutDuplicates(lo123));
+    assertEquals(lo123, CollectionsPlume.withoutDuplicates(lo123123));
+    assertEquals(lo123, CollectionsPlume.withoutDuplicates(lo12223));
+    assertEquals(lo123, CollectionsPlume.withoutDuplicates(lo1123));
+    assertEquals(lo123, CollectionsPlume.withoutDuplicates(lo1233));
+  }
+
+  @SuppressWarnings("ArrayEquals")
+  @Test
+  public void test_withoutDuplicatesComparable() {
+
+    // public static List sortList (List l, Comparator c)
+    // public static <T> List<T> withoutDuplicates(List<T> l) {
+
+    assertEquals(l123, CollectionsPlume.withoutDuplicatesComparable(l123));
+    assertEquals(l123, CollectionsPlume.withoutDuplicatesComparable(l123123));
+    assertEquals(l123, CollectionsPlume.withoutDuplicatesComparable(l12223));
+    assertEquals(l123, CollectionsPlume.withoutDuplicatesComparable(l1123));
+    assertEquals(l123, CollectionsPlume.withoutDuplicatesComparable(l1233));
   }
 
   @Test
@@ -354,7 +384,7 @@ public final class CollectionsPlumeTest {
   }
 
   @Test
-  @SuppressWarnings("lock:methodref.receiver.invalid")
+  @SuppressWarnings("lock:methodref.receiver")
   public void testMapList() {
     List<Object> in = Arrays.asList(new Object[] {1, 2, 3});
     List<Object> out = Arrays.asList(new Object[] {"1", "2", "3"});
@@ -362,7 +392,7 @@ public final class CollectionsPlumeTest {
   }
 
   @Test
-  @SuppressWarnings("lock:methodref.receiver.invalid")
+  @SuppressWarnings("lock:methodref.receiver")
   public void testTransform() {
     List<Object> in = Arrays.asList(new Object[] {1, 2, 3});
     List<Object> out = Arrays.asList(new Object[] {"1", "2", "3"});